+++ MT4Processor/src/com/algotrado/pattern/PTN_0001/PTN_0001_Main.java
@@ -12,8 +12,8 @@ public abstract class PTN_0001_Main implements IPatternState{
 	static protected final String description = ""; // TODO
 	static protected final Integer numberOfStates = Integer.valueOf(2);
 	static protected final String[] parameterStringList = {"The second candle percent"};
-	static protected final Object[] parameterMinValue = {Float.valueOf(0)};
-	static protected final Object[] parameterMaxValue = {Float.valueOf(1)};
+	static protected final Object[] parameterMinValue = {Double.valueOf(0)};
+	static protected final Object[] parameterMaxValue = {Double.valueOf(1)};
 	static protected final Object[] newDataClass = {JapaneseCandleBar.class};
 	static protected final PatternParameterList parametersList = new PTN_0001_Parameters(); // TODO - need to check
